added support for ODROID-N2 hardware platform.

+++ buildroot-external/package/recovery-system/external/overlay/base/bin/fwinstall.sh
@@ -580,7 +580,11 @@ fwinstall()
       echo -ne "OK, "
 
       # on platforms with dedicated boot loaders we have to update them as well.
-      if [[ "${BOOTFS_PLATFORM}" == "tinkerboard" ]] || [[ "${BOOTFS_PLATFORM}" == "ova" ]] || [[ "${BOOTFS_PLATFORM}" == "intelnuc" ]] || [[ "${BOOTFS_PLATFORM}" == "odroid-c4" ]]; then
+      if [[ "${BOOTFS_PLATFORM}" == "tinkerboard" ]] ||
+         [[ "${BOOTFS_PLATFORM}" == "ova" ]] ||
+         [[ "${BOOTFS_PLATFORM}" == "intelnuc" ]] ||
+         [[ "${BOOTFS_PLATFORM}" == "odroid-c4" ]] ||
+         [[ "${BOOTFS_PLATFORM}" == "odroid-n2" ]]; then
         BOOTFS_ROOTDEV="/dev/$(basename "$(dirname "$(readlink "/sys/class/block/${BOOTFS_DEV#/dev/}")")")"
         BOOTFS_START=$(/sbin/fdisk -l "${BOOTFS_ROOTDEV}" | grep FAT32 | head -1 | awk '{ printf $3 }')
         BOOTFS_LOOPROOTDEV=${LOFS_DEV}
@@ -592,10 +596,10 @@ fwinstall()
             echo -ne "(U-Boot)... "
             /bin/dd if="${BOOTFS_LOOPROOTDEV}" of="${BOOTFS_ROOTDEV}" bs=32K count=31 seek=1 skip=1 conv=fsync status=none
             result=$?
-          elif [[ "${BOOTFS_PLATFORM}" == "odroid-c4" ]]; then
-            # ODroid-C4 has U-Boot in seperate boot sector
+          elif [[ "${BOOTFS_PLATFORM}" == "odroid-c4" ]] || [[ "${BOOTFS_PLATFORM}" == "odroid-n2" ]]; then
+            # ODroid-C4/N2 has U-Boot in seperate boot sector
             echo -ne "(U-Boot)... "
-            /bin/dd if="${BOOTFS_LOOPROOTDEV}" of="${BOOTFS_ROOTDEV}" bs=32K count=159 seek=1 skip=1 conv=fsync status=none
+            /bin/dd if="${BOOTFS_LOOPROOTDEV}" of="${BOOTFS_ROOTDEV}" bs=512 count=10239 seek=1 skip=1 conv=fsync status=none
             result=$?
           else
             # x86 RaspberryMatic with GRUB
@@ -748,13 +752,15 @@ touch /tmp/.runningFirmwareUpdate
 
 # fast blink magenta on RPI-RF-MOD or HB-RF-USB/HB-RF-USB-2
 if [[ "${HM_RTC}" == "rx8130" ]] || lsusb | grep -q 0403:6f70 || lsusb | grep -q 10c4:8c07; then
-  echo none  >/sys/class/leds/rpi_rf_mod:green/trigger
-  echo timer >/sys/class/leds/rpi_rf_mod:red/trigger
-  echo timer >/sys/class/leds/rpi_rf_mod:blue/trigger
-  echo 100 >/sys/class/leds/rpi_rf_mod:red/delay_on
-  echo 100 >/sys/class/leds/rpi_rf_mod:red/delay_off
-  echo 100 >/sys/class/leds/rpi_rf_mod:blue/delay_on
-  echo 100 >/sys/class/leds/rpi_rf_mod:blue/delay_off
+  if [[ -e /sys/class/leds/rpi_rf_mod:green/trigger ]]; then
+    echo none  >/sys/class/leds/rpi_rf_mod:green/trigger
+    echo timer >/sys/class/leds/rpi_rf_mod:red/trigger
+    echo timer >/sys/class/leds/rpi_rf_mod:blue/trigger
+    echo 100 >/sys/class/leds/rpi_rf_mod:red/delay_on
+    echo 100 >/sys/class/leds/rpi_rf_mod:red/delay_off
+    echo 100 >/sys/class/leds/rpi_rf_mod:blue/delay_on
+    echo 100 >/sys/class/leds/rpi_rf_mod:blue/delay_off
+  fi
 fi
 
 # if an argument was given (filename of the update file/data)
